Example data in the wide and long formts supported by the evoverse package.

Example data in the wide and long formts supported by the evoverse package.

data(example_input_formats)

Format

4 tibbles, wide/long and mutations/copy number that show the format required to create an evoverse object with function dataset.

Examples

data(example_input_formats)
head(example_input_formats)

#> $wide_mutations
#> # A tibble: 500 x 14
#>    Set6_42 Set6_44     NA NA    NA    NA    NA       NA    NA    NA    NA    NA
#>    <chr>     <int>  <int> <chr> <chr> <chr> <chr> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 chr5     8.98e7 8.98e7 G     A     inte… LYSM… 0         0     0 0         0
#>  2 chrY     1.00e7 1.00e7 G     T     inte… TTTY… 0         0     0 0         0
#>  3 chr12    7.29e7 7.29e7 C     A     intr… TRHDE 0         0     0 0.116    43
#>  4 chr17    2.66e7 2.66e7 G     A     inte… NLK(… 0         0     0 0.127    63
#>  5 chr11    1.14e7 1.14e7 T     G     intr… GALN… 0         0     0 0         0
#>  6 chr1     5.47e7 5.47e7 A     G     intr… MRPL… 0         0     0 0         0
#>  7 chr19    2.45e6 2.45e6 C     G     intr… LMNB2 0         0     0 0         0
#>  8 chr10    8.48e7 8.48e7 A     C     inte… NRG3… 0         0     0 0         0
#>  9 chr3     3.01e7 3.01e7 A     G     inte… RBMS… 0         0     0 0         0
#> 10 chr6     1.00e7 1.00e7 C     T     inte… LOC1… 0.509    53    27 0.513    39
#> # … with 490 more rows, and 2 more variables: NA <dbl>, NA <chr>
#> 
#> $wide_segments
#> # A tibble: 85 x 7
#>    chr      from      to Set6_42.minor Set6_42.Major Set6_44.minor Set6_44.Major
#>    <chr>   <int>   <int>         <int>         <int>         <int>         <int>
#>  1 chr1   1.42e6  2.54e6             1             1             1             1
#>  2 chr1   3.43e6  4.71e7             0             1             0             1
#>  3 chr1   4.87e7  2.48e8             1             1             1             1
#>  4 chr1   2.48e8  2.48e8             0             2             0             2
#>  5 chr2   2.25e5  2.43e8             1             1             1             1
#>  6 chr3   2.91e6  1.97e8             1             1             1             1
#>  7 chr4   8.18e5  1.74e6             1             1             1             1
#>  8 chr4   1.75e6  1.06e7             1             1             1             1
#>  9 chr4   1.35e7  1.91e8             1             1             1             1
#> 10 chr4   1.91e8  1.91e8             1             2             1             2
#> # … with 75 more rows
#> 
#> $long_mutations
#> $long_mutations$Set6_42
#> # A tibble: 175 x 11
#>    chr      from      to ref   alt   region gene         VAF    DP    NV patient
#>    <chr>   <int>   <int> <chr> <chr> <chr>  <chr>      <dbl> <dbl> <dbl> <chr>  
#>  1 chr6   1.00e7  1.00e7 C     T     inter… LOC10050… 0.509     53    27 Set6   
#>  2 chr16  2.06e7  2.06e7 C     T     intro… ACSM2B    0.174     46     8 Set6   
#>  3 chr3   1.80e8  1.80e8 G     A     inter… LOC10050… 0.314     70    22 Set6   
#>  4 chr7   9.77e7  9.77e7 T     A     inter… OCM2(dis… 0.0789    38     3 Set6   
#>  5 chr2   1.39e8  1.39e8 A     C     intro… NXPH2     0.0659    91     6 Set6   
#>  6 chr4   4.92e7  4.92e7 G     T     inter… CWH43(di… 0.167     48     8 Set6   
#>  7 chr20  5.30e7  5.30e7 C     T     inter… PFDN4(di… 0.0721   111     8 Set6   
#>  8 chr15  4.39e7  4.39e7 C     G     intro… CATSPER2  0.0323    93     3 Set6   
#>  9 chr5   1.13e8  1.13e8 T     C     inter… YTHDC2(d… 0.479     94    45 Set6   
#> 10 chr19  4.07e6  4.07e6 G     T     upstr… ZBTB7A    0.366     82    30 Set6   
#> # … with 165 more rows
#> 
#> $long_mutations$Set6_44
#> # A tibble: 146 x 11
#>    chr      from      to ref   alt   region gene         VAF    DP    NV patient
#>    <chr>   <int>   <int> <chr> <chr> <chr>  <chr>      <dbl> <dbl> <dbl> <chr>  
#>  1 chr12  7.29e7  7.29e7 C     A     intro… TRHDE     0.116     43     5 Set6   
#>  2 chr17  2.66e7  2.66e7 G     A     inter… NLK(dist… 0.127     63     8 Set6   
#>  3 chr6   1.00e7  1.00e7 C     T     inter… LOC10050… 0.513     39    20 Set6   
#>  4 chr3   1.80e8  1.80e8 G     A     inter… LOC10050… 0.358     67    24 Set6   
#>  5 chr7   9.77e7  9.77e7 T     A     inter… OCM2(dis… 0.12      25     3 Set6   
#>  6 chr7   5.72e7  5.72e7 G     T     inter… ZNF479(d… 0.0702    57     4 Set6   
#>  7 chr9   1.33e8  1.33e8 T     G     intro… FNBP1     0.0794    63     5 Set6   
#>  8 chr5   1.13e8  1.13e8 T     C     inter… YTHDC2(d… 0.397     63    25 Set6   
#>  9 chrY   1.33e7  1.33e7 C     T     inter… NONE(dis… 0.176     51     9 Set6   
#> 10 chr12  3.79e7  3.79e7 A     G     inter… NONE(dis… 0.0541    74     4 Set6   
#> # … with 136 more rows
#> 
#> 
#> $long_segments
#> $long_segments$Set6_42
#> # A tibble: 85 x 5
#>    chr        from        to minor Major
#>    <chr>     <int>     <int> <int> <int>
#>  1 chr1    1421022   2538349     1     1
#>  2 chr1    3428160  47139103     0     1
#>  3 chr1   48708250 247694949     1     1
#>  4 chr1  247726822 247726822     0     2
#>  5 chr2     224919 242557019     1     1
#>  6 chr3    2908335 197273077     1     1
#>  7 chr4     818440   1739252     1     1
#>  8 chr4    1746125  10599670     1     1
#>  9 chr4   13529990 190894246     1     1
#> 10 chr4  190903688 190904832     1     2
#> # … with 75 more rows
#> 
#> $long_segments$Set6_44
#> # A tibble: 85 x 5
#>    chr        from        to minor Major
#>    <chr>     <int>     <int> <int> <int>
#>  1 chr1    1421022   2538349     1     1
#>  2 chr1    3428160  47139103     0     1
#>  3 chr1   48708250 247694949     1     1
#>  4 chr1  247726822 247726822     0     2
#>  5 chr2     224919 242557019     1     1
#>  6 chr3    2908335 197273077     1     1
#>  7 chr4     818440   1739252     1     1
#>  8 chr4    1746125  10599670     1     1
#>  9 chr4   13529990 190894246     1     1
#> 10 chr4  190903688 190904832     1     2
#> # … with 75 more rows
#> 
#> 
#> $purity
#> Set6_42 Set6_44 
#>     0.8     0.9 
#> 
#> $samples
#> [1] "Set6_42" "Set6_44"
#>