Pipeline to time aneuploidy.

For a list of input mutations, CNA and purity data from a tumour sample, MOBSTER fits are computed to determine proportions of mutations in each aneuploidy state (copy state). Assignments and plots are returned, and an automatic QC pipeline (supervised logistic classifier) is used to flag which fits pass quality check. The trained set for the classifier consists in n = 377 colorectal whole genome samples.

pipeline_chromosome_timing(mutations, cna, purity, karyotypes = c("2:0",
  "2:1", "2:2"), reference = "GRCh38", min_muts = 150,
  min_VAF = 0.05, description = "Chromosomal timing sample",
  N_max = 15000, ...)

Arguments

mutations
Input mutations in the format for package CNAqc.

cna
Absolute Copy Number Alterations in the format for package CNAqc.

purity
Tumour purity, in [0, 1].

karyotypes
Karyotypes that can be timed, expressed as "Major:minor" notation. Default are the following states timeable = c('2:0', '2:1', '2:2').

min_muts
Skip analysing karyotypes with less than min_muts mutations.

...
Parameters passed mobster_fit in package mobster.

Examples

# We use data released with the CNAqc package

cna = CNAqc::example_dataset_CNAqc$cna
mutations = CNAqc::example_dataset_CNAqc$snvs
purity = CNAqc::example_dataset_CNAqc$purity

x = pipeline_chromosome_timing(mutations, cna = cna, purity = purity, auto_setup = 'FAST', N_max = 500)

#>  [ Evoverse Chromosomal timing pipeline ] 
#> 
#> 

#> 

#> ℹ Loading input data

#> 
#> 

#> 

#>                     

#> ── Preparing input for deconvolution ──────────────────────────────────────────────────────────────────

#> ℹ Loading input data

#> 

#>                     

#> ℹ Removing 80 mutations with VAF < 0.05.

#> ℹ Loading input data

#>  [ CNAqc - CNA Quality Check ] 
#> 

#>                     

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Loading input data

#>                     

#> ℹ Drivers are annotated, but 'gene' column is missing, using mutation location.

#> ℹ Loading input data

#>                     

#> ! Missing CCF column from CNA calls, adding CCF = 1 assuming clonal CNA calls.

#> ℹ Loading input data

#>                     

#> ℹ Input n = 13053 mutations for 267 CNA segments (267 clonal, 0 subclonal)

#> ℹ Loading input data

#>                     

#> x n = 172 mutations  cannot be mapped to segments and will be removed.

#> ℹ Loading input data

#> 

#>                     

#> ✓ Mapped n = 12881 mutations to clonal segments (100% of input)

#> ℹ Loading input data

#>                     

#> ── [ CNAqc ]  n = 12881 mutations in 267 segments (267 clonal + 0 subclonal). Genome reference: GRCh38.

#> ℹ Loading input data

#> 
#>    2:2  [n = 7437] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { chr19:15292495:15292496:G:A }
#>    4:2  [n = 1873] ■■■■■■■
#>    3:2  [n = 1612] ■■■■■■
#>    2:1  [n = 1557] ■■■■■■  { chr2:179431633:179431634:C:T }
#>    3:0  [n =  310] ■
#> 

#>                     

#> ℹ Sample Purity: 89% ~ Ploidy: 4.

#> ℹ Loading input data

#> 

#>                     

#> ℹ Mutations annotated have 3 drivers.

#> ℹ Loading input data

#>                     

#> ✓ Loading input data ... done

#> 

#> 

#> 

#> ── MOBSTER clustering mutations with karyotype 2:1 ────────────────────────────────────────────────────

#> 
#> ┌───────────────────────────────────────────────────────┐
#> │                                                       │
#> │   n = 1557 mutations. n > 500 , downsampling input.   │
#> │                                                       │
#> └───────────────────────────────────────────────────────┘
#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Input n = 12881 mutations for 49 CNA segments (49 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> x n = 11324 mutations  cannot be mapped to segments and will be removed.

#> 

#> ✓ Mapped n = 1557 mutations to clonal segments (100% of input)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Input n = 500 mutations for 49 CNA segments (49 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> 

#> ✓ Mapped n = 500 mutations to clonal segments (100% of input)

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 500.

#> ❯ n = 500. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02
#> and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 
#> |===============================================       | 88% ~0 s remaining     |======================================================|100% ~0 s remaining     
#> 

#> ℹ MOBSTER fits completed in 3.4s.

#> 

#> ── [ MOBSTER ] Raw VAF for 2:1 n = 500 with k = 1 Beta(s) without tail ────────────────────────────────

#> ● Clusters: π = 100% [C1], with π > 0.

#> ✖ No tail fit.

#> 

#> ● Beta C1 [n = 500, 100%] with mean = 0.43.

#> ℹ Score(s): NLL = -209.52; ICL = -400.39 (-400.39), H = 0 (0). Fit interrupted by MM in 100 steps.

#> Warning: Unknown or uninitialised column: 'Scale'.

#> Warning: Unknown or uninitialised column: 'Shape'.

#> Warning: Unknown or uninitialised column: 'Scale'.

#> Warning: Unknown or uninitialised column: 'Shape'.

#> 

#> x Raw VAF for 2:1: QC FAIL. p = 0

#> 
#> 
#>         ┌──────────────────────────────────────────────────────────────────────────────────────┐
#>         │                                                                                      │
#>         │   Auto-QC (type 'T') failed, attempting another single run with default parameters   │
#>         │                                                                                      │
#>         └──────────────────────────────────────────────────────────────────────────────────────┘
#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 500.

#> ❯ n = 500. Mixture with k = 1,2,3 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02
#> and n > 10.

#> ❯ Custom fit by Moments-matching in up to 250 steps, with ε = 1e-10 and peaks initialisation.

#> ❯ Scoring (without parallel) 5 x 3 x 2 = 30 models by reICL.

#> 
#> |=======                                               | 13% ~22 s remaining    |=========                                             | 17% ~17 s remaining    |==========                                            | 20% ~15 s remaining    |============                                          | 23% ~20 s remaining    |==============                                        | 27% ~18 s remaining    |================                                      | 30% ~16 s remaining    |==================                                    | 33% ~19 s remaining    |===================                                   | 37% ~17 s remaining    |=====================                                 | 40% ~15 s remaining    |=======================                               | 43% ~16 s remaining    |=========================                             | 47% ~15 s remaining    |===========================                           | 50% ~14 s remaining    |============================                          | 53% ~14 s remaining    |==============================                        | 57% ~13 s remaining    |================================                      | 60% ~11 s remaining    |==================================                    | 63% ~11 s remaining    |====================================                  | 67% ~10 s remaining    |=====================================                 | 70% ~9 s remaining     |=======================================               | 73% ~9 s remaining     |=========================================             | 77% ~8 s remaining     |===========================================           | 80% ~6 s remaining     |=============================================         | 83% ~6 s remaining     |==============================================        | 87% ~5 s remaining     |================================================      | 90% ~3 s remaining     |==================================================    | 93% ~2 s remaining     |====================================================  | 97% ~1 s remaining     |======================================================|100% ~0 s remaining     
#> 

#> ℹ MOBSTER fits completed in 36.2s.

#> 

#> ── [ MOBSTER ] Raw VAF for 2:1 n = 500 with k = 2 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 52% [C2], 43% [C1] and 4% [Tail], with π > 0.

#> ● Tail [n = 18, 4.3%] with alpha = 1.3.

#> ● Beta C1 [n = 217, 43%] with mean = 0.6.

#> ● Beta C2 [n = 265, 52%] with mean = 0.31.

#> ℹ Score(s): NLL = -416.88; ICL = -754.67 (-776.27), H = 23.17 (1.57). Fit converged by MM in 48 steps.

#> 

#> ✓ Raw VAF for 2:1: QC PASS. p = 0.992

#> 
#> 

#> 

#> ── MOBSTER clustering mutations with karyotype 2:2 ────────────────────────────────────────────────────

#> 
#> ┌───────────────────────────────────────────────────────┐
#> │                                                       │
#> │   n = 7437 mutations. n > 500 , downsampling input.   │
#> │                                                       │
#> └───────────────────────────────────────────────────────┘
#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Input n = 12881 mutations for 80 CNA segments (80 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> x n = 5444 mutations  cannot be mapped to segments and will be removed.

#> 

#> ✓ Mapped n = 7437 mutations to clonal segments (100% of input)

#>  [ CNAqc - CNA Quality Check ] 
#> 

#> ℹ Using reference genome coordinates for: GRCh38.

#> ℹ Input n = 500 mutations for 80 CNA segments (80 clonal, 0 subclonal)

#> Warning: [CNAqc] a karyotype column is present in CNA calls, and will be overwritten

#> 

#> ✓ Mapped n = 500 mutations to clonal segments (100% of input)

#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 500.

#> ❯ n = 500. Mixture with k = 1,2 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02
#> and n > 10.

#> ! mobster automatic setup FAST for the analysis.

#> ❯ Scoring (without parallel) 2 x 2 x 2 = 8 models by reICL.

#> 
#> |======================================================|100% ~0 s remaining     
#> 

#> ℹ MOBSTER fits completed in 2.5s.

#> 

#> ── [ MOBSTER ] Raw VAF for 2:2 n = 500 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 86% [C1] and 14% [Tail], with π > 0.

#> ● Tail [n = 64, 14%] with alpha = 0.78.

#> ● Beta C1 [n = 436, 86%] with mean = 0.46.

#> ℹ Score(s): NLL = -611.19; ICL = -1157.82 (-1185.1), H = 27.28 (0). Fit converged by MM in 10 steps.

#> 

#> x Raw VAF for 2:2: QC FAIL. p = 0

#> 
#> 
#>         ┌──────────────────────────────────────────────────────────────────────────────────────┐
#>         │                                                                                      │
#>         │   Auto-QC (type 'T') failed, attempting another single run with default parameters   │
#>         │                                                                                      │
#>         └──────────────────────────────────────────────────────────────────────────────────────┘
#>  [ MOBSTER fit ] 
#> 

#> ✔ Loaded input data, n = 500.

#> ❯ n = 500. Mixture with k = 1,2,3 Beta(s). Pareto tail: TRUE and FALSE. Output clusters with π > 0.02
#> and n > 10.

#> ❯ Custom fit by Moments-matching in up to 250 steps, with ε = 1e-10 and peaks initialisation.

#> ❯ Scoring (without parallel) 5 x 3 x 2 = 30 models by reICL.

#> 
#> |=======                                               | 13% ~24 s remaining    |=========                                             | 17% ~19 s remaining    |==========                                            | 20% ~17 s remaining    |============                                          | 23% ~22 s remaining    |==============                                        | 27% ~19 s remaining    |================                                      | 30% ~20 s remaining    |==================                                    | 33% ~24 s remaining    |===================                                   | 37% ~20 s remaining    |=====================                                 | 40% ~18 s remaining    |=======================                               | 43% ~20 s remaining    |=========================                             | 47% ~17 s remaining    |===========================                           | 50% ~16 s remaining    |============================                          | 53% ~16 s remaining    |==============================                        | 57% ~14 s remaining    |================================                      | 60% ~13 s remaining    |==================================                    | 63% ~13 s remaining    |====================================                  | 67% ~11 s remaining    |=====================================                 | 70% ~10 s remaining    |=======================================               | 73% ~9 s remaining     |=========================================             | 77% ~8 s remaining     |===========================================           | 80% ~7 s remaining     |=============================================         | 83% ~6 s remaining     |==============================================        | 87% ~5 s remaining     |================================================      | 90% ~4 s remaining     |==================================================    | 93% ~3 s remaining     |====================================================  | 97% ~1 s remaining     |======================================================|100% ~0 s remaining     
#> 

#> ℹ MOBSTER fits completed in 39.5s.

#> 

#> ── [ MOBSTER ] Raw VAF for 2:2 n = 500 with k = 1 Beta(s) and a tail ──────────────────────────────────

#> ● Clusters: π = 86% [C1] and 14% [Tail], with π > 0.

#> ● Tail [n = 64, 14%] with alpha = 0.78.

#> ● Beta C1 [n = 436, 86%] with mean = 0.46.

#> ℹ Score(s): NLL = -611.19; ICL = -1157.82 (-1185.1), H = 27.28 (0). Fit converged by MM in 13 steps.

#> 

#> x Raw VAF for 2:2: QC FAIL. p = 0

#> 
#> 

#> 

#> ℹ Pipeline results assembly

#> 

#>                            

#> ✓ Pipeline results assembly ... done

#> 

print(x)

#> ── [ Evoverse ] Chromosomal timing sample Chromosome timing pipeline ──────────────────────────────────

#> 

#> ── [ CNAqc ]  n = 12881 mutations in 267 segments (267 clonal + 0 subclonal). Genome reference: GRCh38.

#> 
#>    2:2  [n = 7437] ■■■■■■■■■■■■■■■■■■■■■■■■■■■  { chr19:15292495:15292496:G:A }
#>    4:2  [n = 1873] ■■■■■■■
#>    3:2  [n = 1612] ■■■■■■
#>    2:1  [n = 1557] ■■■■■■  { chr2:179431633:179431634:C:T }
#>    3:0  [n =  310] ■
#> 

#> ℹ Sample Purity: 89% ~ Ploidy: 4.

#> 

#> ℹ Mutations annotated have 3 drivers.

#> 

#> ──  QC PASS  ─────────────────────────────────────────────────────────────────────── PASS rate (%):  ──

#> # A tibble: 1 x 4
#>   karyotype QC    QC_prob QC_type                
#>   <chr>     <chr>   <dbl> <chr>                  
#> 1 2:1       PASS    0.992 timing_logistic_GEL_CRC
#> 

#> ──  QC PASS  ─────────────────────────────────────────────────────────────────────── FAIL rate (%):  ──

#> # A tibble: 1 x 4
#>   karyotype QC    QC_prob QC_type                
#>   <chr>     <chr>   <dbl> <chr>                  
#> 1 2:2       FAIL        0 timing_logistic_GEL_CRC
#> 
#>   LOG  2020-04-26 02:56:52. evoverse pipeline for chromosome timing: with fits TRUE