Pipeline to perform subclonal deconvolution with MOBSTER.

For a list of input mutations, CNA and purity data from a tumour sample, MOBSTER fits are computed to determine proportions of mutations in each aneuploidy state (copy state). Assignments and plots are returned, and an automatic QC pipeline (supervised logistic classifier) is used to flag which fits pass quality check. The trained set for the classifier consists in n = 377 colorectal whole genome samples.

pipeline_subclonal_deconvolution_CCF(mutations, cna, purity,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2"),
  ccf_method = "ENTROPY", reference = "GRCh38", min_VAF = 0.05,
  min_muts = 50,
  description = "Subclonal deconvolution dataset (MOBSTER + BMix)",
  N_max = 15000, ...)

Arguments

mutations
Input mutations in the format for package CNAqc.

cna
Absolute Copy Number Alterations in the format for package CNAqc.

purity
Tumour purity, in [0, 1].

min_muts
Skip analysing karyotypes with less than min_muts mutations.

...
Parameters passed mobster_fit in package mobster.

timeable
Karyotypes that can be timed, expressed as "Major:minor" notation. Default are the following states timeable = c('2:0', '2:1', '2:2').

Examples

# We use data released with the CNAqc package

cna = CNAqc::example_dataset_CNAqc$cna
mutations = CNAqc::example_dataset_CNAqc$snvs
purity = CNAqc::example_dataset_CNAqc$purity

# We use auto_setup = 'FAST' to speed up the analysis
x = pipeline_subclonal_deconvolution(mutations, cna = cna, purity = purity,  N_max = 1000, auto_setup = 'FAST')

#> Error in pipeline_subclonal_deconvolution(mutations, cna = cna, purity = purity,     N_max = 1000, auto_setup = "FAST"): could not find function "pipeline_subclonal_deconvolution"

print(x)

#> Error in print(x): object 'x' not found