Summary for an evoverse dataset.

The summary reports basic statistics on the stored input values concerning both mutations, segments, and fits if available

# S3 method for mbst_data
summary(x, ...)

Arguments

x
An evoverse object.

...
S3 parameters

Value

Nothing.

Examples

data('example_evoverse')
summary(example_evoverse)

#>  [ mvMOBSTER dataset ] 
#>   Dataset My multi-region MOBSTER dataset 
#>   Samples Set7_55, Set7_57, Set7_59, Set7_62 
#>  Purities 0.88, 0.88, 0.88, 0.8 
#> Mutations N = 57317 
#> 
#> MOBSTER analysis TRUE 
#>   VIBER analysis FALSE 
#> 
#> =-=-=-=-=-=-=-=-
#>  Mutation data
#> =-=-=-=-=-=-=-=-
#> 
#> Depth (DP)  
#> # A tibble: 4 x 6
#>   sample  variable  mean median   min   max
#>   <chr>   <chr>    <dbl>  <dbl> <dbl> <dbl>
#> 1 Set7_55 DP        79.3     80     3  3577
#> 2 Set7_57 DP        77.7     78     3  3889
#> 3 Set7_59 DP        82.6     83     3  3823
#> 4 Set7_62 DP        90.3     90     6  4403
#> 
#> Number of reads with variant (NV)  
#> # A tibble: 4 x 6
#>   sample  variable  mean median   min   max
#>   <chr>   <chr>    <dbl>  <dbl> <dbl> <dbl>
#> 1 Set7_55 NV        9.81      0     0   470
#> 2 Set7_57 NV        7.69      0     0   584
#> 3 Set7_59 NV        8.72      0     0   516
#> 4 Set7_62 NV        7.85      0     0   634
#> 
#> Variant allele frequency (VAF)  
#> # A tibble: 4 x 6
#>   sample  variable  mean median   min   max
#>   <chr>   <chr>    <dbl>  <dbl> <dbl> <dbl>
#> 1 Set7_55 VAF       0.14      0     0  0.76
#> 2 Set7_57 VAF       0.11      0     0  0.76
#> 3 Set7_59 VAF       0.12      0     0  0.73
#> 4 Set7_62 VAF       0.11      0     0  0.76
#> 
#> Annotations gene, vaf_normal, cosmic, function., mutlocation, region 
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  Copy Number Alteration data
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> $Set7_55
#>       CNAqc  n = 47357 mutations for 60 CNA segments (60 clonal, 0 subclonal) 
#>      Purity  88% cellularity 
#>  Karyotypes  47357 (1:1) 
#>  
#>    Peaks QC  FALSE 
#>         CCF  FALSE 
#> 
#> $Set7_57
#>       CNAqc  n = 47357 mutations for 60 CNA segments (60 clonal, 0 subclonal) 
#>      Purity  88% cellularity 
#>  Karyotypes  47357 (1:1) 
#>  
#>    Peaks QC  FALSE 
#>         CCF  FALSE 
#> 
#> $Set7_59
#>       CNAqc  n = 47357 mutations for 60 CNA segments (60 clonal, 0 subclonal) 
#>      Purity  88% cellularity 
#>  Karyotypes  47357 (1:1) 
#>  
#>    Peaks QC  FALSE 
#>         CCF  FALSE 
#> 
#> $Set7_62
#>       CNAqc  n = 47357 mutations for 60 CNA segments (60 clonal, 0 subclonal) 
#>      Purity  80% cellularity 
#>  Karyotypes  47357 (1:1) 
#>  
#>    Peaks QC  FALSE 
#>         CCF  FALSE 
#> 
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  MOBSTER fit for sample Set7_55
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> [ MOBSTER ] N = 20398 points with K = 1 Beta clusters, fit by MM in 10 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 9184 (0.47) 	 Shape = 1.7 
#>   Beta C1 	n = 11214 (0.53) 	 Mean = 0.5
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC  entropy       ICL reduced.entropy     reICL
#>  -17664.43 -35269.33 -35316.87 1598.244 -33671.09               0 -35269.33
#>  size
#>     6
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  MOBSTER fit for sample Set7_57
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> [ MOBSTER ] N = 19305 points with K = 1 Beta clusters, fit by MM in 10 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 11387 (0.61) 	 Shape = 1.4 
#>   Beta C1 	n = 7918 (0.39) 	 Mean = 0.49
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC entropy       ICL reduced.entropy     reICL size
#>  -16489.14 -32919.08 -32966.29 2129.45 -30789.63               0 -32919.08    6
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  MOBSTER fit for sample Set7_59
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> [ MOBSTER ] N = 22575 points with K = 1 Beta clusters, fit by MM in 9 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 14248 (0.65) 	 Shape = 1.8 
#>   Beta C1 	n = 8327 (0.35) 	 Mean = 0.49
#> 
#>   Scores (model selection)  
#>        NLL       BIC      AIC  entropy       ICL reduced.entropy     reICL size
#>  -24866.85 -49673.56 -49721.7 1641.651 -48031.91               0 -49673.56    6
#> 
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#>  MOBSTER fit for sample Set7_62
#> =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
#> [ MOBSTER ] N = 19852 points with K = 1 Beta clusters, fit by MM in 13 steps [CONVERGED] 
#> 
#>   Components (fit)  
#>      Tail	n = 13106 (0.69) 	 Shape = 1.4 
#>   Beta C1 	n = 6746 (0.31) 	 Mean = 0.5
#> 
#>   Scores (model selection)  
#>        NLL       BIC       AIC  entropy       ICL reduced.entropy     reICL
#>  -17069.28 -34079.18 -34126.55 2326.259 -31752.92               0 -34079.18
#>  size
#>     6